# openapi.yaml
openapi: 3.1.0
info:
  title: QuizForge API
  version: 0.5.0-dev
  summary: Asynchronous quiz and assessment tournaments.
  description: |
    QuizForge runs asynchronous quiz tournaments. A tournament opens for a
    window, players enter on their own schedule, and results roll up to a
    leaderboard. There is no synchronous "everyone answers at once" mode, by
    design.

    ## Status of this document

    This describes the API **exactly as it behaves today**.

    The shapes it first described were not all ones worth keeping. Six were
    corrected before anything consumed this document — while doing so was one
    commit rather than a breaking change — and are recorded in
    `docs/api/known-inconsistencies.md` with what they were and why they moved.

    ## Conventions

    - Every identifier is a prefixed `TypeId`: `acc_`, `wsp_`, `trn_`, `att_`,
      `qst_`. Bare UUIDs never appear in a response.
    - Every list endpoint returns `{ "data": [...], "nextCursor": ... }` and
      takes `limit` and `cursor`. Pagination is **keyset**, never offset: a
      cursor names the last row you saw, so a row inserted while you page
      through cannot duplicate or hide another one.
    - Out-of-range parameters are **rejected**, never silently corrected.

    Everything the product does is reachable here: workspaces, members, API
    keys, question banks, questions, tournaments and play.

    ## Questions are immutable

...
    question.

...

...

...

...
    have selected.

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - sessionCookie: []
  - bearerApiKey: []
paths:
...
  /v1/tournaments/{tournamentId}/attempts-summary:
    get:
      tags: [Play]
      operationId: listTournamentAttempts
      summary: Who has played this tournament
      description: |
        The organiser's view: one entry per attempt, newest first, across every
        player.

        Distinct from `GET /v1/tournaments/{id}/standings`, which ranks players.
        This lists attempts, including ones that are still in progress or
        expired and therefore appear on no leaderboard.

        Carries no answers and no question content.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - name: tournamentId
          in: path
          required: true
          description: The tournament.
          schema:
            $ref: '#/components/schemas/TournamentId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: The attempts against this tournament.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptSummaryPage'
        '400':
          $ref: '#/components/responses/InvalidListRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
...
  /v1/tournaments/{tournamentId}/attempts:
    post:
      tags: [Play]
      operationId: startAttempt
      summary: Start an attempt at a tournament
      description: |
        Begins an attempt and freezes its question set immediately, so the
        player is never graded against questions they were not shown, even if
        the bank changes mid-tournament.

        Requires a **session**, not an API key: an attempt is attributed to a
        person, and a key identifies only a workspace.

        Rejected with `400` if the tournament is not currently open, or if the
        account has already used its allowed attempts.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - name: tournamentId
          in: path
          required: true
          description: The tournament to attempt.
          schema:
            $ref: '#/components/schemas/TournamentId'
      responses:
        '201':
          description: The attempt was started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptStarted'
        '400':
          description: |
            The workspace is not in scope, the tournament is not open, or the
            attempt allowance is used up. Carries `code: INVALID_REQUEST`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/CsrfOrPermission'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/attempts/{attemptId}/questions/{position}:
    get:
      tags: [Play]
      operationId: getAttemptQuestion
      summary: Retrieve one question of an attempt
      description: |
        Returns the question at a 1-based position in this attempt's frozen
        set, with options in the order this attempt fixed at creation — so a
        refresh does not reshuffle under the player, while two players still
        see different orders.

        **The correct answer is never included**, in any question type, at any
        point during play.

        Someone else's attempt returns `404` rather than `403`.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/AttemptIdPath'
        - $ref: '#/components/parameters/PositionPath'
      responses:
        '200':
          description: The question, without its answer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayableQuestion'
        '400':
          $ref: '#/components/responses/WorkspaceRequired'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/attempts/{attemptId}/questions/{position}/answer:
    post:
      tags: [Play]
      operationId: answerAttemptQuestion
      summary: Answer one question
      description: |
        Grades the answer on the server and stores the outcome immediately, so
        an attempt survives a disconnect and submission only has to aggregate
        what is already recorded.

        Answering the same position again replaces the stored response, until
        the attempt is closed.

        The response says **whether** the answer was right but never **what**
        the right answer was — otherwise the whole answer key could be
        extracted by submitting one wrong answer per question.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/AttemptIdPath'
        - $ref: '#/components/parameters/PositionPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
      responses:
        '200':
          description: The answer was recorded and graded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerFeedback'
        '400':
          description: |
            The workspace is not in scope, the answer was blank, or the attempt
            has expired or is already closed. Carries `code: INVALID_REQUEST`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/CsrfOrPermission'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/attempts/{attemptId}/submit:
    post:
      tags: [Play]
      operationId: submitAttempt
      summary: Submit an attempt
      description: |
        Closes the attempt and returns its result.

        Submission only aggregates outcomes already stored by each answer, so
        it is naturally idempotent: submitting an already-submitted attempt
        returns the same result rather than an error, and unanswered questions
        count against the score.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/AttemptIdPath'
      responses:
        '200':
          description: The attempt is closed and graded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptResult'
        '400':
          $ref: '#/components/responses/WorkspaceRequired'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/CsrfOrPermission'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/attempts/{attemptId}:
    get:
      tags: [Play]
      operationId: getAttemptResult
      summary: The result of an attempt
      description: |
        The attempt's current score. Readable while the attempt is still in
        progress, in which case it reflects what has been answered so far.
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/AttemptIdPath'
      responses:
        '200':
          description: The attempt's result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptResult'
        '400':
          $ref: '#/components/responses/WorkspaceRequired'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/WorkspaceForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qf_session
      description: |
        Opaque session token issued by `POST /v1/auth/login`, stored only as a
        SHA-256 digest on the server. Mutating requests authenticated this way
        are CSRF-protected and additionally require the `X-XSRF-TOKEN` header.
    bearerApiKey:
      type: http
      scheme: bearer
      description: |
        A workspace API key, sent as `Authorization: Bearer qf_…`. Shown once
        when created and stored only as a SHA-256 digest thereafter.

        Authenticates a workspace rather than a person, so endpoints that
        attribute an action to an account — starting and submitting attempts —
        reject it.
  parameters:
    WorkspaceHeader:
      name: X-QuizForge-Workspace
      in: header
      required: false
      description: |
        The workspace to operate on, as a `wsp_…` identifier.

        Required for every endpoint outside `/v1/auth` when authenticating with
        a session cookie; omitting it fails with `400 INVALID_REQUEST`. With an
        API key the workspace comes from the key and this header is ignored.

        Documented as `required: false` because it is genuinely optional for
        key-authenticated callers, not because it is safe to leave out.

        Naming a workspace you are not a member of returns
        **`403 PERMISSION_DENIED`**: the credential is valid, the workspace is
        not yours. A malformed value is `400 INVALID_REQUEST`.
      schema:
        $ref: '#/components/schemas/WorkspaceId'
    Limit:
      name: limit
      in: query
      required: false
      description: |
        How many items to return, 1–100.

        A value outside that range is `400 INVALID_REQUEST`. It is deliberately
        not clamped: answering a request for 1000 with 100 items is
        indistinguishable from a collection that only holds 100, so a client
        would conclude it had everything when it had a prefix.
      schema:
        type: integer
        format: int32
        default: 25
        minimum: 1
        maximum: 100
    PageCursor:
      name: cursor
      in: query
      required: false
      description: |
        The `nextCursor` from the previous page. Omit for the first page.

        **Treat this as opaque.** It is not a page number, an offset, or an
        identifier, and its encoding is not part of the contract. A value this
        API did not issue is `400 INVALID_CURSOR` rather than being quietly
        treated as "start again".
      schema:
        $ref: '#/components/schemas/CursorParameter'
...
    AttemptIdPath:
      name: attemptId
      in: path
      required: true
      description: |
        The attempt, as a prefixed identifier. Passing an identifier of another
        type — a `trn_…` where an `att_…` belongs — is a `400`, not a
        confusing `404` further along.
      schema:
        $ref: '#/components/schemas/AttemptId'
    PositionPath:
      name: position
      in: path
      required: true
      description: |
        1-based position within this attempt's frozen question set. Not a
        question identifier, and not stable across attempts — two players'
        position 1 are usually different questions.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 1000
  responses:
...
    WorkspaceRequired:
      description: |
        No workspace is in scope. Carries `code: INVALID_REQUEST`.

        Sent when a session-authenticated request omits
        `X-QuizForge-Workspace`. Tenant isolation depends on a workspace being
        selected, so such a request is refused before it reaches a handler
        rather than being served unfiltered.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    WorkspaceForbidden:
      description: |
        The credential is valid but the account is not a member of the workspace
        it named. Carries `code: PERMISSION_DENIED`.

        Distinct from `401` deliberately: re-authenticating cannot fix this, so
        telling the caller they are unauthenticated would send them to the wrong
        remedy.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationRequired:
      description: |
        No valid credential was presented. Carries
        `code: AUTHENTICATION_REQUIRED`.

        Also returned in two cases that are not obviously authentication
        failures:

        - A path that does not exist. Authentication is evaluated before
          routing, so an anonymous caller learns nothing about which paths are
          real.
        It is **not** returned for a workspace you are not a member of — that
        is `403`, because the credential itself is valid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    CsrfOrPermission:
      description: |
        Either the CSRF token was missing on a cookie-authenticated write, or
        the credential lacks the required role.

        The CSRF filter runs **before** authentication, so an unauthenticated
        write returns `403` here rather than `401`. That ordering is
        deliberate; the request is refused either way.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InvalidListRequest:
      description: |
        The request could not be paged. One of:

        - no workspace is in scope — `code: INVALID_REQUEST`
        - `limit` is outside 1–100 — `code: INVALID_REQUEST`
        - `cursor` was not issued by this API — `code: INVALID_CURSOR`

        The three are distinguished by `code`, which is why they share a status
        without sharing a meaning.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: |
        No such object, **or** it belongs to another workspace or another
        account. The two are deliberately indistinguishable, so this endpoint
        cannot be used to discover that an identifier exists.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Internal:
      description: |
        Something failed unexpectedly. The cause is logged server-side and
        never included in the response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      title: Problem
      description: |
        An RFC 9457 Problem Details document. Every error in the API has this
        shape.

        Clients should branch on `code`. It is the only field with a stability
        guarantee: a code that exists will never change meaning or HTTP status,
        and new codes are added rather than existing ones repurposed.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          maxLength: 200
          description: Stable URI identifying the error class.
          examples: ['https://quizforge.dev/errors/not-found']
        title:
          type: string
          maxLength: 100
          description: The error code, repeated as a human-readable title.
          examples: ['NOT_FOUND']
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code, repeated in the body.
          examples: [404]
        detail:
          type: string
          maxLength: 2000
          description: |
            Human-readable explanation. **Not stable** — reworded freely. Never
            parse it.
          examples: ['no such attempt']
        instance:
          type: string
          format: uri-reference
          maxLength: 2000
          description: The path that produced the error.
          examples: ['/v1/attempts/att_01a0136fb3937209a5719e8fb7db26bf']
        code:
          $ref: '#/components/schemas/ErrorCode'
...
    AccountId:
      type: string
      title: AccountId
      description: An account identifier. UUIDv7 rendered as 32 lowercase hex characters with an `acc_` prefix.
      pattern: '^acc_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['acc_01a0136fb3937209a5719e8fb7db26bf']
    WorkspaceId:
      type: string
      title: WorkspaceId
      description: A workspace identifier. UUIDv7 rendered as 32 lowercase hex characters with a `wsp_` prefix.
      pattern: '^wsp_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['wsp_01a0136fb3937209a5719e8fb7db26bf']
    TournamentId:
      type: string
      title: TournamentId
      description: A tournament identifier. UUIDv7 rendered as 32 lowercase hex characters with a `trn_` prefix.
      pattern: '^trn_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['trn_01a0136fb3937209a5719e8fb7db26bf']
    QuestionId:
      type: string
      title: QuestionId
      description: A question identifier. UUIDv7 rendered as 32 lowercase hex characters with a `qst_` prefix.
      pattern: '^qst_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['qst_01a0136fb3937209a5719e8fb7db26bf']
    Cursor:
      type: [string, 'null']
      title: Cursor
      description: |
        Pass this back as `cursor` to get the next page. `null` means there is
        no next page — stop, rather than asking for one you already know is
        empty.

        **Opaque.** Not a page number, not an offset, not an identifier. Its
        encoding is not part of the contract and may change; a cursor issued by
        an older version will be rejected rather than misread.
      maxLength: 500
      examples: ['djE6MDFhMDEzNmZiMzkzNzIwOWE1NzE5ZThmYjdkYjI2YmY']
    CursorParameter:
      type: string
      title: CursorParameter
      description: An opaque cursor previously returned as `nextCursor`.
      maxLength: 500
      examples: ['djE6MDFhMDEzNmZiMzkzNzIwOWE1NzE5ZThmYjdkYjI2YmY']
...
    AttemptId:
      type: string
      title: AttemptId
      description: An attempt identifier. UUIDv7 rendered as 32 lowercase hex characters with an `att_` prefix.
      pattern: '^att_[0-9a-f]{32}$'
      maxLength: 36
      examples: ['att_01a0136fb3937209a5719e8fb7db26bf']
...
    AttemptStarted:
      type: object
      title: AttemptStarted
      description: What a player needs to begin. Carries no question content.
      required: [id, questions]
      properties:
        id:
          $ref: '#/components/schemas/AttemptId'
        questions:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
          description: |
            The **number** of questions in this attempt, not a list of them.
            Frozen now, so later changes to the bank cannot alter this
            attempt's denominator.
        expiresAt:
          type: [string, 'null']
          format: date-time
          maxLength: 40
          description: |
            When this attempt stops accepting answers. `null` when the
            tournament sets no time limit.
      examples:
        - id: att_01a0136fb3937209a5719e8fb7db26bf
          questions: 10
          expiresAt: '2026-08-18T10:25:00Z'
    QuestionType:
      type: string
      title: QuestionType
      enum: [SINGLE_CHOICE, MULTI_CHOICE, TRUE_FALSE, NUMERIC, SHORT_TEXT]
      examples: ['SINGLE_CHOICE']
    PlayableQuestion:
      type: object
      title: PlayableQuestion
      description: |
        A question as shown to a player. **Never carries the correct answer**,
        for any type, at any point during the attempt.
      required: [questionId, type, prompt, options]
      properties:
        questionId:
          $ref: '#/components/schemas/QuestionId'
        type:
          $ref: '#/components/schemas/QuestionType'
        prompt:
          type: string
          maxLength: 4000
        options:
          type: array
          maxItems: 26
          description: |
            Selectable options, in the order this attempt fixed at creation.
            Empty for `NUMERIC` and `SHORT_TEXT`, which are free-entry.
          items:
            type: string
            maxLength: 1000
      examples:
        - questionId: qst_01a0136fb3937209a5719e8fb7db26bf
          type: SINGLE_CHOICE
          prompt: What is the capital of France?
          options: [Paris, Lyon, Marseille, Toulouse]
    AnswerRequest:
      type: object
      title: AnswerRequest
      required: [answer]
      properties:
        answer:
          type: string
          minLength: 1
          maxLength: 4000
          description: |
            The player's answer, always a string regardless of question type.

            For `MULTI_CHOICE`, the selected options separated by `|`. For
            `NUMERIC`, the number as text, compared within the question's
            tolerance. Blank is rejected as `400`, not graded as incorrect.
      examples:
        - answer: Paris
    AnswerFeedback:
      type: object
      title: AnswerFeedback
      description: |
        What a player learns after answering. Deliberately never includes the
        correct answer, even when the given answer was wrong.
      required: [position, correct, answered, total, hasNext]
      properties:
        position:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
          description: The position just answered.
        correct:
          type: boolean
        answered:
          type: integer
          format: int32
          minimum: 0
          maximum: 1000
          description: How many distinct positions have been answered so far.
        total:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
        hasNext:
          type: boolean
          description: Whether a further position exists in this attempt.
      examples:
        - position: 1
          correct: true
          answered: 1
          total: 10
          hasNext: true
    AttemptState:
      type: string
      title: AttemptState
      description: |
        `EXPIRED` means the time limit passed; such an attempt is still scored
        on what was answered and still appears on the leaderboard.
      enum: [STARTED, GRADED, EXPIRED]
      examples: ['GRADED']
    AttemptResult:
      type: object
      title: AttemptResult
      description: The outcome of an attempt.
      required: [attemptId, state, score, outOf, percentage]
      properties:
        attemptId:
          $ref: '#/components/schemas/AttemptId'
          description: |
            The same string `POST /v1/tournaments/{id}/attempts` returned, so a
            client can use the identifier it was given without translating it.
        state:
          $ref: '#/components/schemas/AttemptState'
        score:
          type: integer
          format: int32
          minimum: 0
          maximum: 1000
          description: Questions answered correctly.
        outOf:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
          description: |
            The denominator frozen when the attempt started. Unanswered
            questions count against the player.
        percentage:
          type: number
          format: double
          minimum: 0
          maximum: 100
      examples:
        - attemptId: att_01a0136fb3937209a5719e8fb7db26bf
          state: GRADED
          score: 8
          outOf: 10
          percentage: 80
...
    AttemptSummary:
      type: object
      title: AttemptSummary
      description: |
        One player's attempt, as the organiser sees it. Carries no answers and
        no question content.
      required: [id, accountId, state, startedAt, outOf]
      properties:
        id:
          $ref: '#/components/schemas/AttemptId'
        accountId:
          $ref: '#/components/schemas/AccountId'
        state:
          $ref: '#/components/schemas/AttemptState'
        startedAt:
          type: string
          format: date-time
          maxLength: 40
        submittedAt:
          type: [string, 'null']
          format: date-time
          maxLength: 40
          description: When the attempt closed, or `null` while it is still open.
        score:
          type: [integer, 'null']
          format: int32
          minimum: 0
          maximum: 1000
          description: |
            Correct answers so far. Present while in progress too — an attempt
            is graded as it goes, not at submission.
        outOf:
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
          description: |
            The denominator frozen when the attempt started, so it is unaffected
            by later changes to the bank.
      examples:
        - id: att_01a0136fb3937209a5719e8fb7db26bf
          accountId: acc_01a0136fb3937209a5719e8fb7db26bf
          state: GRADED
          startedAt: '2026-09-01T10:00:00Z'
          submittedAt: '2026-09-01T10:06:00Z'
          score: 8
          outOf: 10
    AttemptSummaryPage:
      type: object
      title: AttemptSummaryPage
      required: [data]
      properties:
        data:
          type: array
          maxItems: 10000
          items:
            $ref: '#/components/schemas/AttemptSummary'
        nextCursor:
          $ref: '#/components/schemas/Cursor'